Refactor e2e suite code
